Delete old Id implementation and replace with new global one. Treat clippy warnings
