+++ src/simulation/mod.rs
@@ -1,7 +1,6 @@
 pub mod config;
 pub mod controller;
 pub mod id;
-mod id2;
 pub mod io;
 pub mod logging;
 pub mod messaging;
